refactor(pdf-service): moveFillOutFunction one level up
