+++ app/services/pdf/beratungshilfe/sections/__test__/C_einkommen.test.ts
@@ -1,7 +1,7 @@
 import { type BeratungshilfeFormularContext } from "~/flows/beratungshilfeFormular";
 import { getBeratungshilfeParameters } from "~/services/pdf/beratungshilfe";
 import { fillEinkommen } from "~/services/pdf/beratungshilfe/sections/C_einkommen";
-import { pdfFillReducer } from "~/services/pdf/prozesskostenhilfe/fillOutFunction";
+import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 
 describe("C_einkommen", () => {
   it("should fill partner einkommen pdf fields when context is partner", () => {
