+++ app/services/pdf/prozesskostenhilfe/J_belastungen.ts
@@ -1,5 +1,5 @@
 import { finanzielleAngabeEinkuenfteGuards as einkuenfteGuards } from "~/flows/prozesskostenhilfeFormular/finanzielleAngaben/einkuenfte/guards";
-import type { PkhPdfFillFunction } from "./fillOutFunction";
+import type { PkhPdfFillFunction } from ".";
 import { SEE_IN_ATTACHMENT_DESCRIPTION } from "../beratungshilfe/sections/E_unterhalt";
 import { ausgabenSituationMapping } from "../beratungshilfe/sections/G_ausgaben";
